requirements.txt
streamlit==1.49.1
python-dotenv==1.1.1
google-generativeai==0.8.5
requests==2.32.5
numpy
plotly==6.3.0
faiss-cpu==1.12.0
sentence-transformers==3.0.1
torch>=2.0.0

# for predicting
h5py==3.14.0
xarray==2024.7.0
tensorflow==2.20.0
scipy==1.13.1
netCDF4==1.7.2
h5netcdf==1.6.4
arm_pyart==1.19.2

# for reading
supabase
ijson

